broadcastHashJoin fix - util class renamed
